Align Fumadocs setup
- .source handling: Stopped tracking .source in git (git rm --cached), left it on disk only as generated output, and kept it in .gitignore.
- Runtime fix for fumadocs-mdx:collections/server: Switched back to the default .source outDir (no custom outDir or Turbopack alias) so the dev server resolves the module on Windows.
- Config updates for default .source: next.config.ts uses createMDX() with no options; tsconfig.json maps fumadocs-mdx:collections/* to .source/*; package.json has postinstall: "fumadocs-mdx" and types:check: "fumadocs-mdx && next typegen && tsc --noEmit" (no custom outDir args).
- Docs/comments: README and source.config.ts now describe MDX output as generated into .source/ (gitignored).
- app/layout.tsx metadata: Replaced ogTitle/ogDescription with title/description and use them for root metadata, Open Graph, and Twitter card (single source of truth).
