+++ .gitignore
@@ -1,7 +1,7 @@
 # deps
 /node_modules
 
-# generated content
+# generated by fumadocs-mdx (do not commit)
 .source
 
 # test & build
