--- .source/dynamic.ts
@@ -1,8 +1,0 @@
-// @ts-nocheck
-import { dynamic } from 'fumadocs-mdx/runtime/dynamic';
-import * as Config from '../source.config';
-
-const create = await dynamic<typeof Config, import("fumadocs-mdx/runtime/types").InternalTypeConfig & {
-  DocData: {
-  }
-}>(Config, {"configPath":"source.config.ts","environment":"next","outDir":".source"}, {"doc":{"passthroughs":["extractedReferences"]}});
