--- .source/source.config.mjs
@@ -1,23 +1,0 @@
-// source.config.ts
-import { defineConfig, defineDocs, frontmatterSchema, metaSchema } from "fumadocs-mdx/config";
-var docs = defineDocs({
-  dir: "content/docs",
-  docs: {
-    schema: frontmatterSchema,
-    postprocess: {
-      includeProcessedMarkdown: true
-    }
-  },
-  meta: {
-    schema: metaSchema
-  }
-});
-var source_config_default = defineConfig({
-  mdxOptions: {
-    // MDX options
-  }
-});
-export {
-  source_config_default as default,
-  docs
-};
