--- next.config.mjs
@@ -1,12 +1,0 @@
-import { createMDX } from 'fumadocs-mdx/next';
-
-const withMDX = createMDX();
-
-/** @type {import('next').NextConfig} */
-const config = {
-  reactStrictMode: true,
-  output: 'export',
-  images: { unoptimized: true },
-};
-
-export default withMDX(config);
