+++ next.config.ts
@@ -1,0 +1,13 @@
+import { createMDX } from 'fumadocs-mdx/next';
+import type { NextConfig } from 'next';
+
+// Use default .source outDir so fumadocs-mdx/Next resolve fumadocs-mdx:collections/* at runtime (custom outDir + alias fails on Windows dev server)
+const withMDX = createMDX();
+
+const config: NextConfig = {
+  reactStrictMode: true,
+  output: 'export',
+  images: { unoptimized: true },
+};
+
+export default withMDX(config);
