+++ postcss.config.mjs
@@ -1,5 +1,8 @@
-export default {
+/** @type {import('postcss-load-config').Config} */
+const config = {
   plugins: {
     '@tailwindcss/postcss': {},
   },
 };
+
+export default config;
