+++ source.config.ts
@@ -1,3 +1,8 @@
+/**
+ * Fumadocs config (TypeScript, canonical).
+ * This file lives at the repo root. Fumadocs-mdx generates MDX collections into .source/
+ * on build/postinstall (no .source directory).
+ */
 import { defineConfig, defineDocs, frontmatterSchema, metaSchema } from 'fumadocs-mdx/config';
 
 // You can customise Zod schemas for frontmatter and `meta.json` here
